requirements.txt
asgiref==3.5.1
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djongo==1.3.3
dnspython==2.2.1
importlib-metadata==4.11.3
Markdown==3.3.6
mongoengine==0.24.1
pymongo==3.12.1
pytz==2022.1
sqlparse==0.2.4
tzdata==2022.1
zipp==3.8.0
